audio-recognition-system/requirements.txt
# MVP版音声認識・翻訳・Google Docs出力システム
# 動作確認済み依存関係ファイル
# 
# 生成環境: macOS 14.5.0, Python 3.13.x, Apple Silicon
# 生成日時: 2025年1月25日
# 動作確認: 完全テスト済み（音声認識・翻訳・Google Docs出力）
#
# 使用方法:
#   pip install -r requirements.txt
#
# 注意事項:
#   - macOS以外の環境では一部パッケージ（mlx-*）がインストールされない場合があります
#   - その場合は自動的にフォールバック機能が動作します

accelerate==1.7.0
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-speech==2.33.0
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mlx==0.25.2
mlx-lm==0.24.1
mlx-whisper==0.4.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
openai-whisper==20240930
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
zipp==3.22.0
